requirement.txt
streamlit
requests
beautifulsoup4
langchain
langchain-google-genai
python-dotenv
